[CRTOOL] Moving utility and content page related files into their own folders
