+++ teachers_digital_platform/crtool/src/js/components/pages/SurveyPageContainer.js
@@ -7,8 +7,8 @@ import ContentHighCriterionPage from "./content/ContentHighCriterionPage";
 import ContentElementarySummaryPage from "./content/ContentElementarySummaryPage";
 import ContentHighSummaryPage from "./content/ContentHighSummaryPage";
 import ContentMiddleSummaryPage from "./content/ContentMiddleSummaryPage";
-import UtilityCriterionPage from "./UtilityCriterionPage";
-import UtilitySummaryPage from "./UtilitySummaryPage";
+import UtilityCriterionPage from "./utility/UtilityCriterionPage";
+import UtilitySummaryPage from "./utility/UtilitySummaryPage";
 import QualityCriterionPage from "./QualityCriterionPage";
 import QualitySummaryPage from "./QualitySummaryPage";
 import EfficacyCriterionPage from "./EfficacyCriterionPage";
