+++ teachers_digital_platform/crtool/src/js/components/pages/utility/UtilityCriterionRow.js
